+++ packages/expo-eas-client/ios/EASClient/EASClientModule.swift
@@ -6,8 +6,6 @@ public class EASClientModule: Module {
   public func definition() -> ModuleDefinition {
     Name("EASClient")
 
-    Constants([
-      "clientID": EASClientID.uuid().uuidString
-    ])
+    Constant("clientID") { EASClientID.uuid().uuidString }
   }
 }